Clearly indicate compile-time constants
When managing data structures, it is very
important to know which sizes are set at
compile time and which are variables.
I've changed the prefix of the Global constants
to "c_" instead of "g_" to make this clear.
I've also declared them as typed enums, which
from Kokkos developer experience is the
most portable way to ensure the compiler
doesn't ever store them in memory.